requirements.txt
imageio==2.19.3
imutils==0.5.4
networkx==2.8.3
numpy==1.22.4
opencv-contrib-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
pytesseract==0.3.9
PyWavelets==1.3.0
scikit-image==0.19.2
scipy==1.8.1
tifffile==2022.5.4
pyjwt[crypto]
ipython
argparse==1.4.0
requests==2.25.1
pyyaml==5.3.1